Fixed runtime bugs in the demo app and auto-set up server on port 6380
