+++ demo-redis-app/utils/mockCommandData.js
