+++ demo-redis-app/utils/mockCommandSettings.js
@@ -12,7 +12,7 @@ module.exports = {
   STRING_TIME_TO_LIVE: 120,
   LIST_LPUSH_FREQUENCY: 1000,
   LIST_LRANGE_FREQUENCY: 1000,
-  LIST_RPOP_FREQUENCY: 5000,
+  LIST_LPOP_FREQUENCY: 5000,
   LIST_TIME_TO_LIVE: 120,
   SET_SADD_FREQUENCY: 1000,
   SET_SMEMBERS_FREQUENCY: 1000,
